fix(Plugin, UnitTests): Fix read single record bug + adjust first discovered schema for unit tests
resolves DATAINT-1528

+++ PluginFileReader/API/Utility/LoadDirectoryFilesIntoDb.cs
@@ -1,6 +1,5 @@
 using System.Collections.Generic;
 using System.Linq;
-using System.Text.RegularExpressions;
 using Naveego.Sdk.Logging;
 using Newtonsoft.Json;
 using PluginFileReader.API.Factory;
@@ -13,12 +12,13 @@ namespace PluginFileReader.API.Utility
     {
         public static void LoadDirectoryFilesIntoDb(
             IImportExportFactory factory, SqlDatabaseConnection conn, RootPathObject rootPath,
-            string tableName, string schemaName, List<string> paths, bool downloadToLocal, long recordLimit = long.MaxValue, int fileLimit = int.MaxValue
-            )
+            string tableName, string schemaName, List<string> paths, bool downloadToLocal, long recordLimit = long.MaxValue, int fileLimit = int.MaxValue,
+            bool deleteAllOnStart = true)
         {
             Logger.Info($"Loading files:\n {JsonConvert.SerializeObject(paths, Formatting.Indented)}");
             
-            DeleteDirectoryFilesFromDb(conn, tableName, schemaName, factory, rootPath, paths);
+            if (deleteAllOnStart)
+                DeleteDirectoryFilesFromDb(conn, tableName, schemaName, factory, rootPath, paths);
 
             foreach (var path in paths.Take(fileLimit))
             {
